Store signatures for every past version of a file
This is needed for resolving conflicts. If a patch depends on some
older version of file (which is true only for conflicts) we need to
know how to apply this patch. Thus we have to store signatures for each
version.

In future we can do some prunning on signatures so they take less
space.